% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_OTU_counts.R
\name{build_OTU_counts}
\alias{build_OTU_counts}
\title{Batch - build_OTU_counts analysis of many comparisons}
\usage{
build_OTU_counts(
  biom = NULL,
  sample_table = NULL,
  taxa = NULL,
  verbose = FALSE
)
}
\arguments{
\item{biom}{A "phyloseq" object with included groups
to be analysed. For format specifications see ?phyloseq E.g.
accessible as "build_OTU_counts$otu". Groups are used to automate colouring of
samples. Default = NULL}

\item{sample_table}{A sample table to map groups to corresponding samples.  Default = NULL}

\item{taxa}{taxanomic rank to match ranks. Default = NULL}

\item{verbose}{Verbosity ON/OFF. Default=FALSE}
}
\value{
A list of all OTU comparisons conducted.
}
\description{
Given a summarized experiment generated using buildSummarized()
this function will automatically perform differential expression (DE)
analysis for all possible groups using 3 different methods 1) EdgeR, 2) Voom
and 3) DEseq2. It will also output 10x diagnostic plots automatically, if the
plotting options are selected (see ?diag_plots for more details).
}
