% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTU_plots.R
\name{OTU_plots}
\alias{OTU_plots}
\title{Batch - OTU_plots}
\usage{
OTU_plots(build_OTU_counts_output = NULL, verbose = FALSE)
}
\arguments{
\item{build_OTU_counts_output}{A "phyloseq" object with included groups
to be analysed. For format specifications see ?phyloseq E.g.
accessible as "build_OTU_counts$otu". Groups are used to automate colouring of
samples. Default = NULL
after DE analysis from merged results. Options: TRUE, FALSE. Default = TRUE}

\item{verbose}{Verbosity ON/OFF. Default=FALSE}
}
\value{
A list of all OTU comparisons conducted.
}
\description{
Given a summarized experiment generated using buildSummarized()
this function will automatically perform differential expression (DE)
analysis for all possible groups using 3 different methods 1) EdgeR, 2) Voom
and 3) DEseq2. It will also output 10x diagnostic plots automatically, if the
plotting options are selected (see ?diag_plots for more details).
}
